requirements.txt
sphinx==2.1.2
pylint==2.3.1
numpy==1.16.4
astropy==3.2.1
pillow==6.1.0
scipy==1.3.0
astropy==3.2.1
opencv-python==4.1.1.26
matplotlib==3.1.0
requests_cache==0.5.2
dippykit==2.0.2
ATpy==0.9.7
wget==3.2
